created settings-data file and  split settings tests into 2 spec files
